tests/requirements.txt
beautifulsoup4
pytest
pyyaml
requests
